chore(optimization): some refactors and optimizations
