+++ src/components/datatable/components/TableHeader/types/TableHeader.types.ts
@@ -1,8 +1,6 @@
 import { SearchFieldProps } from "../../SearchFields/types/SearchForm.types";
-import { TableOperationProps } from "../../TableOperations/types/TableOperations.types";
 
 export type TableHeaderProps<FormObjectType, RecordType> = SearchFieldProps<
   FormObjectType,
   RecordType
-> &
-  TableOperationProps;
+>;
