--- src/components/datatable/components/TableOperations/styles/TableOperationsWrapper.tsx
@@ -1,5 +1,0 @@
-import styled from "styled-components";
-
-const TableOperationsWrapper = styled.div``;
-
-export default TableOperationsWrapper;
