--- src/components/datatable/components/TableOperations/types/TableOperations.types.ts
@@ -1,5 +1,0 @@
-import { ReactNode } from "react";
-
-export interface TableOperationProps {
-  tableOperations?: ReactNode;
-}
